+++ src/module/ModuleValue.cpp
@@ -10,9 +10,11 @@
 // ============================================
 
 ModuleValue::ModuleValue(const std::string& id, const std::string& name, QueryPeriod period,
-    const std::string& field)
+    const std::string& field, std::optional<int> min_value, std::optional<int> max_value)
     : id_(id)
     , name_(name)
     , query_period_(period)
-    , field_(field) {
+    , field_(field)
+    , min_value_(min_value)
+    , max_value_(max_value) {
 }